Require lineify in PATH.
We've been relying on lineify being in a particular place on the
filesystem; this is ridiculously silly.

Instead, require the `lineify` executable to be somewhere in the PATH,
which is how it'll be in normal usage.  Currently, this requires the
user to create a symlink manually, since lineify hasn't been gemified
(see xiongchiamiov/lineify#4), but that's at least an API that should be
there in the future.